More compiler warnings, mostly from NetBSD

+++ ntpq/ntpq.c
@@ -3378,7 +3378,7 @@ static void list_md_fn(const EVP_MD *m, const char *from, const char *to, void *
     /* Lowercase names aren't accepted by keytype_from_text in ssl_init.c */
 
     for( cp = name; *cp; cp++ ) {
-	if( islower(*cp) )
+	if( islower((int)*cp) )
 	    return;
     }
     len = (cp - name) + 1;
